+++ merlin/examples/generator.py
@@ -48,11 +48,19 @@ LOG = logging.getLogger(__name__)
 
 EXAMPLES_DIR = os.path.join(os.path.dirname(os.path.abspath(__file__)), "workflows")
 
+# TODO modify the example command to eliminate redundancy
+# - e.g. running `merlin example flux_local` will produce the same output
+#   as running `merlin example flux_par` or `merlin example flux_par_restart`.
+#   This should just be `merlin example flux`.
+# - restart and restart delay should be one example
+# - feature demo and remote feature demo should be one example
+# - all openfoam examples should just be under one openfoam label
+
 
 def gather_example_dirs():
     """Get all the example directories"""
     result = {}
-    for directory in os.listdir(EXAMPLES_DIR):
+    for directory in sorted(os.listdir(EXAMPLES_DIR)):
         result[directory] = directory
     return result
 
@@ -82,7 +90,7 @@ def list_examples():
     for example_dir in gather_example_dirs():
         directory = os.path.join(os.path.join(EXAMPLES_DIR, example_dir), "")
         specs = glob.glob(directory + "*.yaml")
-        for spec in specs:
+        for spec in sorted(specs):
             if "template" in spec:
                 continue
             with open(spec) as f:  # pylint: disable=C0103