+++ tests/fixtures/status.py
@@ -14,6 +14,9 @@ import yaml
 from tests.unit.study.status_test_files import status_test_variables
 
 
+# pylint: disable=redefined-outer-name
+
+
 @pytest.fixture(scope="session")
 def status_testing_dir(temp_output_dir: str) -> str:
     """
@@ -29,8 +32,8 @@ def status_testing_dir(temp_output_dir: str) -> str:
     return testing_dir
 
 
-@pytest.fixture(scope="session")
-def status_empty_file(status_testing_dir: str) -> str:  # pylint: disable=W0621
+@pytest.fixture(scope="class")
+def status_empty_file(status_testing_dir: str) -> str:
     """
     A pytest fixture to create an empty status file.